changed database from sqlite to postgresql

+++ db/migrate/20150611200456_add_location_to_bike_racks.rb
@@ -1,6 +1,6 @@
 class AddLocationToBikeRacks < ActiveRecord::Migration
   def change
-    add_column :bike_racks, :latitude, :double
-    add_column :bike_racks, :longitude, :double
+    add_column :bike_racks, :latitude, :float
+    add_column :bike_racks, :longitude, :float
   end
 end
